+++ shenyu-examples/shenyu-examples-springmvc/src/main/http/http-test-api.http
@@ -16,7 +16,7 @@
 #
 
 ### shengyu getway proxy orderSave
-POST http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/order/save
+POST http://localhost:9195/mvc/order/save
 Accept: application/json
 Content-Type: application/json
 
@@ -26,28 +26,28 @@ Content-Type: application/json
 }
 
 ### shengyu getway proxy orderSave
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/order/findById?id=123
+GET http://localhost:9195/mvc/order/findById?id=123
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy path
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/order/path/123/hahah
+GET http://localhost:9195/mvc/order/path/123/hahah
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy path
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/order/path/123/name
+GET http://localhost:9195/mvc/order/path/123/name
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy oauth2
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/order/oauth2/test
+GET http://localhost:9195/mvc/order/oauth2/test
 Accept: application/json
 Content-Type: application/json
 Authorization:AuthorizationContext
 
 ### shengyu getway proxy payment
-POST http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/payment
+POST http://localhost:9195/mvc/test/payment
 Accept: application/json
 Content-Type: application/json
 
@@ -57,31 +57,31 @@ Content-Type: application/json
 }
 
 ### shengyu getway proxy findByUserId
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/findByUserId?userId=123
+GET http://localhost:9195/mvc/test/findByUserId?userId=123
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy path
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/path/123?name=haha
+GET http://localhost:9195/mvc/test/path/123?name=haha
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy path
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/path/123/name
+GET http://localhost:9195/mvc/test/path/123/name
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy waf pass
-POST http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/waf/pass
+POST http://localhost:9195/mvc/test/waf/pass
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy waf deny
-POST http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/waf/deny
+POST http://localhost:9195/mvc/test/waf/deny
 Accept: application/json
 Content-Type: application/json
 
 ### shengyu getway proxy request parameter pass
-GET http://localhost:9195/mvc/shenyu_examples_springmvc_war_exploded/test/request/parameter/pass?requestParameter=test
+GET http://localhost:9195/mvc/test/request/parameter/pass?requestParameter=test
 Accept: application/json
 Content-Type: application/json
