--- shenyu-examples/shenyu-examples-springmvc/src/main/java/com/apache/shenyu/springmvc/dto/UserDTO.java
@@ -1,82 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.apache.shenyu.springmvc.dto;
-
-import java.util.StringJoiner;
-
-/**
- * The type User dto.
- */
-public class UserDTO {
-
-    /**
-     * user id
-     */
-    private String userId;
-
-    /**
-     * username
-     */
-    private String userName;
-
-    /**
-     * Get userId.
-     *
-     * @return userId
-     */
-    public String getUserId() {
-        return userId;
-    }
-
-    /**
-     * Set userId.
-     *
-     * @param userId userId
-     */
-    public void setUserId(final String userId) {
-        this.userId = userId;
-    }
-
-    /**
-     * Get userName.
-     *
-     * @return userName
-     */
-    public String getUserName() {
-        return userName;
-    }
-
-    /**
-     * Set userName.
-     *
-     * @param userName userName
-     */
-    public void setUserName(final String userName) {
-        this.userName = userName;
-    }
-
-    @Override
-    public String toString() {
-        return new StringJoiner(", ", UserDTO.class.getSimpleName() + "[", "]")
-                .add("userId='" + userId + "'")
-                .add("userName='" + userName + "'")
-                .toString();
-    }
-
-
-}
