+++ shenyu-examples/shenyu-examples-springmvc/src/main/java/org/apache/shenyu/examples/springmvc/result/ResultBean.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.apache.shenyu.springmvc.result;
+package org.apache.shenyu.examples.springmvc.result;
 
 /**
  * ResultBean.
