+++ shenyu-examples/shenyu-examples-springmvc/src/main/resources/context/mvc.xml
@@ -21,7 +21,7 @@
        xmlns:context="http://www.springframework.org/schema/context"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
     <mvc:annotation-driven/>
-    <context:component-scan base-package="com.apache.shenyu.springmvc.controller" use-default-filters="false">
+    <context:component-scan base-package="org.apache.shenyu.examples.springmvc.controller" use-default-filters="false">
         <context:include-filter type="annotation" expression="org.springframework.web.bind.annotation.RestController"/>
     </context:component-scan>
     <mvc:default-servlet-handler/>
